+++ src/modules/moralis/moralis.service.ts
@@ -1,5 +1,7 @@
 import { HttpService, Injectable, Logger } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { InjectQueue } from '@nestjs/bull';
+import { Queue } from 'bull';
 import { In, Repository } from 'typeorm';
 import * as crypto from 'crypto';
 import Moralis from 'moralis/node';
@@ -7,7 +9,6 @@ import { customAlphabet } from 'nanoid';
 import Downloader from 'nodejs-file-downloader';
 
 import { AppConfig } from '../configuration/configuration.service';
-import { QueueService } from '../queue/queue.service';
 
 import { S3Service } from '../file-storage/s3.service';
 import { Nft, NftSource } from '../nft/domain/nft.entity';
@@ -25,27 +26,23 @@ import {
   ImageUriFormatNotSupportedError,
   TokenAssertAddressNotSupportedError,
   OpenSeaNftImageSupportedError,
+  ImageDownloadError,
 } from './service/exceptions';
 import { FileSystemService } from '../file-system/file-system.service';
 import { NftValidator } from './service/nft-validator';
-
-const MORALIS_NEW_NFT_QUEUE = 'MORALIS_NEW_NFT_QUEUE';
-const OPENSEA_RINKEBY_API_URL = 'https://rinkeby-api.opensea.io/api/v1/asset';
-const OPENSEA_ETH_API_URL = 'https://api.opensea.io/api/v1/asset';
-const MORALIS_IPFS_SERVER_URL = 'https://ipfs.moralis.io:2053/ipfs/';
-
-enum MetaDataApiCallType {
-  TOKEN_URI = 1,
-  OPENSEA = 2,
-}
-
+import {
+  MORALIS_NEW_NFT_QUEUE,
+  MORALIS_IPFS_SERVER_URL,
+  PROCESS_MORALIS_TOKEN_JOB,
+  OPENSEA_NFT_QUEUE,
+  PROCCESS_OPENSEA_NFT,
+} from './constants';
 @Injectable()
 export class MoralisService {
   private logger = new Logger(MoralisService.name);
 
   constructor(
     private readonly config: AppConfig,
-    private readonly queue: QueueService,
     private httpService: HttpService,
     private s3Service: S3Service,
     private fileSystemService: FileSystemService,
@@ -60,6 +57,10 @@ export class MoralisService {
     private monitoredNftsRepository: Repository<MonitoredNfts>,
     @InjectRepository(MoralisLog)
     private moralisLogRepository: Repository<MoralisLog>,
+    @InjectQueue(MORALIS_NEW_NFT_QUEUE)
+    private readonly moralisNftQueue: Queue,
+    @InjectQueue(OPENSEA_NFT_QUEUE)
+    private readonly openseaNftQueue: Queue,
   ) {}
 
   private routeIpfsUrlImageIpfs(url: string) {
@@ -80,7 +81,6 @@ export class MoralisService {
     const take = 100;
     let keepLooping = true;
     let skip = 0;
-    let idsToDelete = [];
     this.logger.log('retry');
 
     while (keepLooping) {
@@ -94,11 +94,8 @@ export class MoralisService {
         if (log.token && log.token.symbol) {
           try {
             await this.processToken(log.token);
-            idsToDelete.push(log.id);
-            if (idsToDelete.length > 100) {
-              await this.moralisLogRepository.delete({ id: In(idsToDelete) });
-              idsToDelete = [];
-            }
+            await this.moralisLogRepository.delete({ id: log.id });
+            console.log('remove successfully', log.token);
           } catch (error) {
             this.logger.error(`retry ${log.id} ${error}`);
             continue;
@@ -109,7 +106,10 @@ export class MoralisService {
       if (logs.length < take) keepLooping = false;
       skip = skip + take;
     }
-    await this.moralisLogRepository.delete({ id: In(idsToDelete) });
+  }
+
+  async sleep(ms) {
+    return new Promise((resolve) => setTimeout(resolve, ms));
   }
 
   async getHistory(start: number, end: number) {
@@ -117,7 +117,7 @@ export class MoralisService {
     const pageSize = 500;
     let keepLooping = true;
     while (keepLooping) {
-      console.log(`Fetch NFTs from ${skip} to ${skip + pageSize}`);
+      this.logger.log(`Fetch NFTs from ${skip} to ${skip + pageSize}`);
       const params = {
         start: start,
         end: end,
@@ -128,65 +128,58 @@ export class MoralisService {
       try {
         results = await Moralis.Cloud.run('fetchEthNFTOwners', params);
       } catch (err) {
-        console.log(err);
+        this.logger.log(err);
       }
 
       if (results.length < pageSize) keepLooping = false;
       for (const token of results) {
-        this.addNewNFT(token);
+        await this.addNewNFT(token);
+      }
+      while (true) {
+        const queueCount = await this.moralisNftQueue.count();
+        const queueCount1 = await this.openseaNftQueue.count();
+        if (queueCount < 1000 && queueCount1 < 10000) break;
+        this.logger.debug(`Current position is ${skip}, moralis=${queueCount}, opensea=${queueCount1}`);
+        await this.sleep(20000);
       }
       skip += pageSize;
     }
   }
 
   async onModuleInit() {
-    Moralis.serverURL = this.config.values.moralis.serverUrl;
-    Moralis.masterKey = this.config.values.moralis.masterKey;
-    Moralis.initialize(this.config.values.moralis.applicationId);
-    this.queue.initQueue(MORALIS_NEW_NFT_QUEUE, this.moralisNewNFTOwnerHandler, 1);
+    await Moralis.start({
+      serverUrl: this.config.values.moralis.serverUrl,
+      appId: this.config.values.moralis.applicationId,
+      masterKey: this.config.values.moralis.masterKey,
+    });
   }
 
   addNewUserToWatchAddress = async (address: string) => {
-    Moralis.Cloud.run('watchEthAddress', {
-      address,
-      chainId: this.config.values.ethereum.ethereumNetwork === 'rinkeby' ? '0x4' : '0x1',
-      sync_historical: true,
-    });
+    await Moralis.Cloud.run(
+      'watchEthAddress',
+      {
+        address,
+        chainId: this.config.values.ethereum.ethereumNetwork === 'rinkeby' ? '0x4' : '0x1',
+        sync_historical: true,
+      },
+      { useMasterKey: true },
+    );
   };
 
   async addNewNFT(token) {
-    this.queue.pushToQueue(MORALIS_NEW_NFT_QUEUE, { token });
+    this.logger.log('Add new nft to the moralis queue', token);
+    await this.moralisNftQueue.add(PROCESS_MORALIS_TOKEN_JOB, { token });
   }
 
-  moralisNewNFTOwnerHandler = async (input: any, cb: any) => {
+  public async moralisNewNFTOwnerHandler(input: any) {
     const { token }: { token: MoralisNft } = input;
-    this.logger.log(token);
 
     try {
       await this.processToken(token);
-    } catch (error) {
-      if (
-        error instanceof NftMissingAttributesError ||
-        error instanceof TokenUriFormatNotSupportedError ||
-        error instanceof ImageUriFormatNotSupportedError ||
-        error instanceof OpenSeaNftImageSupportedError ||
-        error instanceof TokenAssertAddressNotSupportedError
-      ) {
-        const newMoralisLog = this.moralisLogRepository.create();
-        newMoralisLog.name = error.name;
-        newMoralisLog.token = token;
-        await this.moralisLogRepository.save(newMoralisLog);
-      } else if (!(error instanceof SkippedUniverseNftError)) {
-        const newMoralisLog = this.moralisLogRepository.create();
-        newMoralisLog.name = 'Unknown';
-        newMoralisLog.token = token;
-        await this.moralisLogRepository.save(newMoralisLog);
-      }
-      console.log(error);
+    } catch (err) {
+      this.logger.error(err);
     }
-
-    cb(null, true);
-  };
+  }
 
   private async processToken(token: MoralisNft) {
     this.nftValidator.checkNftHasAllAttributes(token);
@@ -205,12 +198,43 @@ export class MoralisService {
       // create only non-Universe NFTs
       await this.checkNftIsNotUniverseContract(token.token_address);
       await this.checkNftIsNotCoreUniverseContract(token.token_address);
-      existingNft = await this.createNewNft(token, existingCollection);
+
+      try {
+        existingNft = await this.createNewNft(token, existingCollection);
+        this.logger.log(`Scraped token successfully {${token.token_address} - ${token.token_id}}`);
+      } catch (error) {
+        await this.openseaNftQueue.add(PROCCESS_OPENSEA_NFT, {
+          tokenAddress: token.token_address,
+          amount: token.amount,
+          tokenId: token.token_id,
+          collectionId: existingCollection.id,
+        });
+        if (
+          error instanceof NftMissingAttributesError ||
+          error instanceof ImageDownloadError ||
+          error instanceof TokenUriFormatNotSupportedError ||
+          error instanceof ImageUriFormatNotSupportedError ||
+          error instanceof OpenSeaNftImageSupportedError ||
+          error instanceof TokenAssertAddressNotSupportedError
+        ) {
+          const newMoralisLog = this.moralisLogRepository.create();
+          newMoralisLog.name = error.name;
+          newMoralisLog.token = token;
+          await this.moralisLogRepository.save(newMoralisLog);
+        } else if (!(error instanceof SkippedUniverseNftError)) {
+          const newMoralisLog = this.moralisLogRepository.create();
+          newMoralisLog.name = 'Unknown';
+          newMoralisLog.token = token;
+          await this.moralisLogRepository.save(newMoralisLog);
+        }
+      }
     }
   }
 
   private getTokenUri(token: MoralisNft) {
-    if (token.token_uri.includes('0x{id}')) {
+    if (token.token_uri.includes('opensea')) {
+      throw new TokenUriFormatNotSupportedError('opensea nft');
+    } else if (token.token_uri.includes('0x{id}')) {
       return token.token_uri.replace('0x{id}', token.token_id); //Cyber Girls
     } else if (token.token_uri.includes('{id}')) {
       return token.token_uri.replace('{id}', token.token_id); //Fluf erc-1155
@@ -219,16 +243,12 @@ export class MoralisService {
     }
   }
 
-  private async parseNewNFTMetaData(
-    existingNft: Nft,
-    token: MoralisNft,
-    numberOfEditions,
-    _apiCallType: MetaDataApiCallType = MetaDataApiCallType.TOKEN_URI,
-  ) {
+  private async parseNewNFTMetaData(existingNft: Nft, token: MoralisNft, numberOfEditions) {
     if (!!token.token_uri) {
-      const { apiCallType, metadata } = await this.getTokenUriMetaData(existingNft.tokenUri, token, _apiCallType);
+      const metadata = await this.getTokenUriMetaData(existingNft.tokenUri);
       existingNft.name = metadata.name;
       existingNft.description = metadata.description;
+      existingNft.external_link = metadata.external_url;
 
       if (metadata.isImageOnIPFS()) {
         const ipfsImageUrl = this.routeIpfsUrlImageIpfs(metadata.getImage());
@@ -243,11 +263,7 @@ export class MoralisService {
         try {
           await downloader.download();
         } catch (err) {
-          if (apiCallType !== MetaDataApiCallType.OPENSEA) {
-            return await this.parseNewNFTMetaData(existingNft, token, numberOfEditions, MetaDataApiCallType.OPENSEA);
-          } else {
-            throw new OpenSeaNftImageSupportedError();
-          }
+          throw new ImageDownloadError(ipfsImageUrl);
         }
         const s3Result = await this.s3Service.uploadDocument(downloadPath, filename);
         existingNft.artworkType = metadata.getFileExtension() && metadata.getFileExtension().split('.').slice(-1)[0];
@@ -275,11 +291,7 @@ export class MoralisService {
         try {
           await downloader.download();
         } catch (err) {
-          if (apiCallType !== MetaDataApiCallType.OPENSEA) {
-            return await this.parseNewNFTMetaData(existingNft, token, numberOfEditions, MetaDataApiCallType.OPENSEA);
-          } else {
-            throw new OpenSeaNftImageSupportedError();
-          }
+          throw new ImageDownloadError(metadata.getImage());
         }
         const s3Result = await this.s3Service.uploadDocument(downloadPath, filename);
         existingNft.artworkType = metadata.getFileExtension() && metadata.getFileExtension().split('.').slice(-1)[0];
@@ -297,11 +309,7 @@ export class MoralisService {
         try {
           decoded = this.fileSystemService.decodeBase64(metadata.getImage());
         } catch (err) {
-          if (apiCallType !== MetaDataApiCallType.OPENSEA) {
-            return await this.parseNewNFTMetaData(existingNft, token, numberOfEditions, MetaDataApiCallType.OPENSEA);
-          } else {
-            throw new OpenSeaNftImageSupportedError();
-          }
+          throw new OpenSeaNftImageSupportedError();
         }
         const extension = decoded.type.split('/')[1].split('+')[0];
         const filename = `${await this.generateRandomHash()}.${extension}`;
@@ -392,62 +400,27 @@ export class MoralisService {
     }
   }
 
-  private async getTokenMetaDataWithOpenSeaAPI(token: MoralisNft) {
-    if (token.token_id == undefined || token.token_address == undefined) {
-      throw new TokenAssertAddressNotSupportedError();
-    } else {
-      const openSeaApiUri =
-        this.config.values.ethereum.ethereumNetwork === 'rinkeby' ? OPENSEA_RINKEBY_API_URL : OPENSEA_ETH_API_URL;
-      const headers =
-        this.config.values.ethereum.ethereumNetwork === 'rinkeby'
-          ? {}
-          : { 'X-API-KEY': this.config.values.opensea.apiKey };
-      const { data } = await this.httpService
-        .get(`${openSeaApiUri}/${token.token_address}/${token.token_id}`, {
-          headers,
-        })
-        .toPromise();
-
-      const metadata = new StandardNftMetadata(data);
-      return metadata;
-    }
-  }
-
-  private async getTokenUriMetaData(
-    tokenUri: string,
-    token: MoralisNft,
-    apiCallType: MetaDataApiCallType = MetaDataApiCallType.TOKEN_URI,
-  ) {
-    if (apiCallType === MetaDataApiCallType.OPENSEA) {
-      return {
-        apiCallType: MetaDataApiCallType.OPENSEA,
-        metadata: await this.getTokenMetaDataWithOpenSeaAPI(token),
-      };
-    }
+  private async getTokenUriMetaData(tokenUri) {
     try {
       if (tokenUri.startsWith('ipfs')) {
         const normalizedTokenUri = this.routeIpfsUrlToMoralisIpfs(tokenUri);
         const { data } = await this.httpService.get(normalizedTokenUri).toPromise();
         const metadata = new StandardNftMetadata(data);
-        return { apiCallType: MetaDataApiCallType.TOKEN_URI, metadata };
+        return metadata;
       } else if (tokenUri.startsWith('http')) {
         const normalizedTokenUri = tokenUri;
         const { data } = await this.httpService.get(normalizedTokenUri).toPromise();
         const metadata = new StandardNftMetadata(data);
-        return { apiCallType: MetaDataApiCallType.TOKEN_URI, metadata };
+        return metadata;
       } else if (tokenUri.startsWith('data:application/json;base64,')) {
         const data = this.nftValidator.parseBase64TokenUri(tokenUri);
         const metadata = new StandardNftMetadata(data);
-        return { apiCallType: MetaDataApiCallType.TOKEN_URI, metadata };
-      } else {
-        throw new TokenUriFormatNotSupportedError(tokenUri);
+        return metadata;
       }
-    } catch {
-      return {
-        apiCallType: MetaDataApiCallType.OPENSEA,
-        metadata: await this.getTokenMetaDataWithOpenSeaAPI(token),
-      };
+    } catch (err) {
+      throw new TokenUriFormatNotSupportedError(tokenUri);
     }
+    throw new TokenUriFormatNotSupportedError(tokenUri);
   }
 
   private async generateRandomHash(length = 24): Promise<string> {