+++ src/modules/nft/entrypoints/dto.ts
@@ -2,6 +2,7 @@ import {
   ArrayMaxSize,
   ArrayMinSize,
   IsArray,
+  IsInt,
   IsNumber,
   IsNumberString,
   IsOptional,
@@ -24,6 +25,51 @@ export class EditCollectionBody {
     required: false,
   })
   description?: string;
+
+  @IsString()
+  @IsOptional()
+  @ApiProperty({
+    example: 'https://site.io',
+    description: 'The site of the collection',
+    required: false,
+  })
+  siteLink?: string;
+
+  @IsString()
+  @IsOptional()
+  @ApiProperty({
+    example: 'universe.xyz',
+    description: 'The discord link of the collection',
+    required: false,
+  })
+  discordLink?: string;
+
+  @IsString()
+  @IsOptional()
+  @ApiProperty({
+    example: 'universeXYZ',
+    description: 'Instagram profile of the collection',
+    required: false,
+  })
+  instagramLink?: string;
+
+  @IsString()
+  @IsOptional()
+  @ApiProperty({
+    example: 'universeXYZ',
+    description: 'The medium profile of the collection',
+    required: false,
+  })
+  mediumLink?: string;
+
+  @IsString()
+  @IsOptional()
+  @ApiProperty({
+    example: 'universeXYZ',
+    description: 'The telegram profile of the collection',
+    required: false,
+  })
+  telegramLink?: string;
 }
 
 export class SaveNftBody {
@@ -408,6 +454,56 @@ export class CreateCollectionBody {
     required: false,
   })
   description?: string;
+
+  @IsString()
+  @IsOptional()
+  @Length(1, 1024)
+  @ApiProperty({
+    example: 'https://site.io',
+    description: 'The site of the collection',
+    required: false,
+  })
+  siteLink?: string;
+
+  @IsString()
+  @IsOptional()
+  @Length(1, 1024)
+  @ApiProperty({
+    example: 'universe.xyz',
+    description: 'The discord link of the collection',
+    required: false,
+  })
+  discordLink?: string;
+
+  @IsString()
+  @IsOptional()
+  @Length(1, 1024)
+  @ApiProperty({
+    example: 'universeXYZ',
+    description: 'Instagram profile of the collection',
+    required: false,
+  })
+  instagramLink?: string;
+
+  @IsString()
+  @IsOptional()
+  @Length(1, 1024)
+  @ApiProperty({
+    example: 'universeXYZ',
+    description: 'The medium profile of the collection',
+    required: false,
+  })
+  mediumLink?: string;
+
+  @IsString()
+  @IsOptional()
+  @Length(1, 1024)
+  @ApiProperty({
+    example: 'universeXYZ',
+    description: 'The telegram profile of the collection',
+    required: false,
+  })
+  telegramLink?: string;
 }
 
 export class EditMintingCollectionParams {
@@ -438,6 +534,74 @@ export class GetUserNftsParams {
     required: true,
   })
   username: string;
+}
+
+export class GetUserNftsQueryParams {
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The offset for getting nfts',
+    example: '8',
+  })
+  offset: number;
+
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The amount of editions to get',
+    example: '1',
+  })
+  limit: number;
+
+  @IsOptional()
+  @ApiProperty({
+    description: 'The collection ids to filter by',
+    example: '12,3',
+  })
+  collections: string;
+
+  @IsOptional()
+  @IsString()
+  @ApiProperty({
+    description: 'NFT name',
+  })
+  name: string;
+}
+
+export class GetSavedNftsParams {
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The offset for getting nfts',
+    example: '8',
+  })
+  offset: number;
+
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The amount of editions to get',
+    example: '1',
+  })
+  limit: number;
+}
+
+export class GetMyNftsParams {
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The offset for getting nfts',
+    example: '8',
+  })
+  offset: number;
+
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The amount of editions to get',
+    example: '1',
+  })
+  limit: number;
 }
 
 export class EditMintingCollectionBody {
@@ -480,6 +644,14 @@ export class EditMintingNftBody {
     required: false,
   })
   txHash: string;
+
+  @IsInt()
+  @ApiProperty({
+    example: 10,
+    description: 'Number of actual editions minted',
+    required: false,
+  })
+  numberOfEditions: number;
 }
 
 export class GetNftParams {
@@ -495,6 +667,42 @@ export class GetNftParams {
     example: '1',
   })
   tokenId: number;
+}
+
+export class GetMyCollectionsPendingParams {
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The offset for getting nfts',
+    example: '8',
+  })
+  offset: number;
+
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The amount of editions to get',
+    example: '1',
+  })
+  limit: number;
+}
+
+export class GetMyNftsPendingParams {
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The offset for getting nfts',
+    example: '8',
+  })
+  offset: number;
+
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The amount of editions to get',
+    example: '1',
+  })
+  limit: number;
 }
 export class GetMyNftsAvailabilityParams {
   @IsNumberString()
@@ -520,14 +728,6 @@ export class GetMyNftsAvailabilityParams {
     example: '3',
   })
   size: number;
-
-  @IsNumberString()
-  @IsOptional()
-  @ApiProperty({
-    description: 'The id of the auction',
-    example: '1',
-  })
-  auctionId: number;
 }
 
 export class GetMyCollectionsParams {
@@ -537,6 +737,40 @@ export class GetMyCollectionsParams {
     example: 'true',
   })
   mintable: string;
+
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The offset for getting nfts',
+    example: '8',
+  })
+  offset: number;
+
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The amount of editions to get',
+    example: '1',
+  })
+  limit: number;
+}
+
+export class GetMyCollectionsTabParams {
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The offset for getting nfts',
+    example: '8',
+  })
+  offset: number;
+
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The amount of editions to get',
+    example: '1',
+  })
+  limit: number;
 }
 
 export class GetCollectionQueryParams {
@@ -564,4 +798,36 @@ export class GetCollectionQueryParams {
     example: '1',
   })
   limit: number;
+}
+
+export class GetMyNftsPageParams {
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The offset for getting nfts',
+    example: '8',
+  })
+  offset: number;
+
+  @IsNumberString()
+  @IsOptional()
+  @ApiProperty({
+    description: 'The amount of editions to get',
+    example: '1',
+  })
+  limit: number;
+
+  @IsOptional()
+  @ApiProperty({
+    description: 'The collection ids to filter by',
+    example: '12,3',
+  })
+  collections: string;
+
+  @IsOptional()
+  @IsString()
+  @ApiProperty({
+    description: 'NFT name',
+  })
+  name: string;
 }